requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
sentence-transformers
faiss-cpu
python-dotenv
pydantic
scikit-learn
pytest
httpx